+++ main.c
@@ -5,28 +5,33 @@
 #include <argp.h>
 #include <error.h>
 #include <stdlib.h>
+#include <stdbool.h>
 
 // Constants for argument parsing (flags)
 const char *argp_program_version = "grep-in-c 0.0.2";
 static char doc[] = "A grep clone";
 
 /* Description of accepted arguments */
-static char args_doc[] = "";
+static char args_doc[] = "[QUIET?][PARTIAL?] FILE PATTERN";
 
 /* Options the program understands */
 static struct argp_option options[] = {
     {"quiet", 'q', 0, OPTION_ARG_OPTIONAL, "Don't print app banner"},
     {"file", 'f', "FILE", 0, "File to which the pattern should be applied to"},
     {"pattern", 'p', "PATTERN", 0, "Pattern to search for in <file>"},
+    {"partial", 'i', 0, 0, "Partial match pattern against input file"},
+    {"debug", 'd', 0, 0, "Show debug output"},
     { 0 },
 };
 
-/* Used by main to communicate with parse_opt */
+/* Used by main to communicate with `parse_opt` */
 struct arguments {
-    char *args[4];
+    char *args[3]; // Change this every time you expect positional arguments
     int silent;
     char *file;
     char *pattern;
+    bool partial;
+    bool debug;
 };
 
 /* Parse a single argument */
@@ -46,7 +51,14 @@ static error_t parse_opt(int key, char *arg, struct argp_state *state)
         case 'p':
             arguments->pattern = arg;
             break;
+        case 'i':
+            arguments->partial = true;
+            break;
+        case 'd':
+            arguments->debug = true;
+            break;
         default:
+            //printf("k: %c\n", key);
             return ARGP_ERR_UNKNOWN;
     }
     return 0;
@@ -62,6 +74,7 @@ char DATA[256];
 
 // Function signatures
 void compare(FILE* f, char* string);
+void partial(FILE* f, char* string);
 
 
 // Main
@@ -73,23 +86,30 @@ int main(int argc, char** argv) {
     arguments.silent = 0;
     arguments.file = "";
     arguments.pattern = "";
+    arguments.partial = false;
+    arguments.debug = false;
 
     /* Parse our arguments; every option seen by parse_opt will be reflected in arguments. */
     argp_parse(&argp, argc, argv, 0, 0, &arguments);
 
-    /* Debugging print */
-    //printf("arguments.silent: %s\narguments.file: %s\narguments.pattern: %s\n", arguments.silent ? "yes" : "no", arguments.file, arguments.pattern);
+    if (arguments.debug == 1) {
+        /* Debugging print */
+        printf("arguments.silent: %s\narguments.file: %s\narguments.pattern: "
+             "%s\narguments.partial: %s\narguments.debug: %s\n",
+             arguments.silent ? "yes" : "no", arguments.file, arguments.pattern,
+             arguments.partial ? "yes" : "no", arguments.debug ? "yes" : "no");
+    }
 
     if (arguments.silent != 1) {
-      printf("%s: The new grep!\n", P_NAME);
+        printf("%s: The new grep!\n", P_NAME);
     }
     
     if (argc < 3) {
-        error(2, 0, "[ERROR]: Not enough arguments.\nSYNTAX: %s", SYNTAX);
+        error(2, 0, "[ERROR]: Not enough arguments.\n%s", SYNTAX);
     }
 
     if (strcmp(arguments.file, "") == 0) {
-        error(ENOENT, ENOENT, "[ERROR]: Empty file");
+        error(ENOENT, ENOENT, "[ERROR]: No file supplied");
     }
 
     if (strcmp(arguments.pattern, "") == 0) {
@@ -103,7 +123,7 @@ int main(int argc, char** argv) {
 
 	// Check that file exists
     if (fptr == NULL) {
-      error(2, 0, "[ERROR]: %s", ENOENT);
+      error(ENOENT, ENOENT, "[ERROR]: ");
     }
 
 	// stat structure
@@ -122,9 +142,14 @@ int main(int argc, char** argv) {
         exit(2);
     }
     
-	// Compare input string (argv[2]) to lines within the file pointer to find exact matches
-    compare(fptr, string);
-    
+    if (arguments.partial) {
+        // Find substring within file pointer (partial matches)
+        partial(fptr, string);
+    } else {
+      // Compare input string (argv[2]) to lines within the file pointer to find
+      // exact matches
+      compare(fptr, string);
+    }
     
     // close file when done
     fclose(fptr);
@@ -146,6 +171,22 @@ void compare(FILE* f, char* string) {
             printf("[MATCH]: %s\n", DATA);
         } else {
             printf("-\n");
+        }
+    }
+}
+
+// Find partial matches between supplied pattern and file
+void partial(FILE* f, char* string) {
+    // TODO: Add colorizing matching pattern
+    while (fgets(DATA, sizeof(DATA), f) != NULL) {
+
+        // Debug print
+        //printf("[fgets()]: %s\n", DATA);
+
+        if (strstr(DATA, string) != NULL) {
+            printf("[MATCH]: %s\n", DATA);
+        } else {
+            fprintf(stderr, "[ERROR]: No match found!\n");
         }
     }
 }