+++ RELEASES.md
@@ -1,4 +1,5 @@
 | Version | Changes |
 |  :---:  |---------|
-| 0.0.1   | "Initial Release" |
-| 0.0.2   | "Added CLI flag parsing, more optimized code" |
+| 0.0.1   | Initial Release |
+| 0.0.2   | Added CLI flag parsing, more optimized code |
+| 0.0.3   | Added partial string match, debug output |